pyproject.toml
[project]
name = "llm-mcp-python-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.25.0",
    "numpy>=2.4.1",
    "openai>=2.15.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml==6.0.3",
    "rich>=14.2.0",
]
